pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fineye"
version = "0.1.0"
description = "Personal finance tracker with voice-driven expense capture, pandas/CSV ledger, and AI categorization."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "FinEye" }]
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "jinja2>=3.1",
  "python-dotenv>=1.0",
  "python-multipart>=0.0.9",
  "pandas>=2.2",
  "openpyxl>=3.1",
  "chardet>=5.2",
  "rapidfuzz>=3.9",
  "ulid-py>=1.1",
  "httpx>=0.27",
  "PyYAML>=6.0",
  "python-dateutil>=2.9",
  "gspread>=6.0",
  "google-auth>=2.30",
  "google-auth-oauthlib>=1.2",
  "psycopg2-binary>=2.9",
  "itsdangerous>=2.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "hypothesis>=6.100",
  "ruff>=0.5",
  "mypy>=1.10",
  "types-python-dateutil",
  "types-PyYAML",
  "respx>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "SIM"]
ignore = ["E501", "B008", "SIM105", "SIM117"]

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["app.services.*", "app.storage.*"]
strict = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]
